Add Turnstile-protected waitlist entry flow
- Add @marsidev/react-turnstile and a small server-side verifier helper.
- Require Turnstile on /api/waitlist/join when both site and secret keys
  are configured; fail closed if only one side of the config is present.
- Add Turnstile UI to the sold-out public waitlist form and reset the
  widget after success/failure.
- Add admin waitlist links from the events table and event edit dialog.
- Fix waitlist advisory lock SQL casts so Postgres resolves the intended
  pg_advisory_lock(int, int) overload.
- Trigger marketing-site event revalidation from event create/update/delete.
- Add staff-facing waitlist guide docs and ignore local .vercel metadata.

Made-with: Cursor

+++ src/lib/services/waitlist/allocateWaitlistOffers.ts
@@ -4,10 +4,10 @@ import { computeAvailability } from '../ticketing/availability';
 
 const OFFER_DURATION_MINUTES = 120;
 
-// Stable advisory lock key derived from eventDateId.
-// pg_advisory_lock takes a bigint; we use the eventDateId directly
-// (prefixed with a namespace constant to avoid collisions with other locks).
-const LOCK_NAMESPACE = 834719; // arbitrary constant
+// Stable advisory lock key: Postgres exposes pg_advisory_lock(int, int) or
+// pg_advisory_lock(bigint) — not (bigint, bigint). Cast in SQL so parameters
+// are not inferred as bigint pairs.
+const LOCK_NAMESPACE = 834719; // arbitrary constant (fits int4)
 
 interface AllocationResult {
   offersCreated: number;
@@ -43,7 +43,7 @@ export async function allocateWaitlistOffers(params: {
   // This serialises concurrent allocation runs so two callers can't
   // both read the same availability snapshot and over-offer.
   await prisma.$executeRawUnsafe(
-    `SELECT pg_advisory_lock($1, $2)`,
+    `SELECT pg_advisory_lock($1::int, $2::int)`,
     LOCK_NAMESPACE,
     eventDateId
   );
@@ -52,7 +52,7 @@ export async function allocateWaitlistOffers(params: {
     return await runAllocation(params);
   } finally {
     await prisma.$executeRawUnsafe(
-      `SELECT pg_advisory_unlock($1, $2)`,
+      `SELECT pg_advisory_unlock($1::int, $2::int)`,
       LOCK_NAMESPACE,
       eventDateId
     );
